Using --enablerepo='lsst-ts*' to specifiy the repo instead of defining daily, bleed or release. Added a doc statement to the waitUntil loop.
Stripping the .preNNN string from the sal_version when defining the docker container EnvVar.

Added a properies block to disable concurrent builds and define the buildDiscarder properties.

Added a new parameter, build_type, that defines from which repo to pull the RPM. The flag value is set in the build stage based on the parameter value.